Response mechanism enhanced. User and Customer actions separated. CRUD almost impl. pending update user and change admin status

+++ src/main/java/workshop/repositories/DatabaseCustomerRepository.java
@@ -3,6 +3,6 @@ package workshop.repositories;
 import org.springframework.data.jpa.repository.JpaRepository;
 import workshop.model.customer.Customer;
 
-public interface DatabaseCustomerRepository extends JpaRepository<Customer, Integer> {
+public interface DatabaseCustomerRepository extends JpaRepository<Customer, Long> {
 
 }
